+++ src/components/Admin/NewFeatured.vue
@@ -29,11 +29,11 @@
 
 <script setup>
 import { computed, ref, watch } from 'vue';
-import { CREATE_FEATURED, GET_PIN } from '../config/constants'
+import { CREATE_FEATURED, GET_PIN } from '../../config/constants'
 import { useMutation, useApolloClient } from '@vue/apollo-composable';
 import Spinner from '../components/Layout/Spinner.vue';
-import callAdminServer from '../utils/callAdminServer';
-import { useWalletStore } from '../stores/wallet';
+import callAdminServer from '../../utils/callAdminServer';
+import { useWalletStore } from '../../stores/wallet';
 const walletStore = useWalletStore()
 const siteID = import.meta.env.VITE_WEBSITE_ID;
 const adminServerUrl = import.meta.env.VITE_ADMIN_SERVER;
